Continued to build combat engine
Jammed out a few hours of coding
Implemented the functionality of CombatAlpha state on GameManager & functionality of CombatBeta state
Added in methods on Combatant script for
returning type, atk type, rolling priority, rolling attack values and taking damage
Created TypeComparison script which checks the type of an ATK and the type of a DEF then sets a value for a modifier. Uses a lot of nested if-statements, feels like it could be cleaner.
Added methods for setting AlphaCombat, BetaCombat and DeclareWinner states in GameManager
Added variables for 'alpha combatant' and 'beta combatant' (ie, going first and going second)
Added method CalculatePriority to GameManager which calls the RollPriority methods on both the alpha combatant and beta combatant
Added method CheckWinner to GameManager which checks if a defender's HP is below 0 and returns true or false
Added in debugs to track where everything is going and tested if the loop is working
So far working, with player selecting a fighter and then the player fighter and bot fighter exchanging hits until one is at 0hp before going to DeclareWinner state.
Next step: modify GameManager to include an int for PlayerFightersAlive and BotFightersAlive, so that the DeclareWinner state can check if one or the other is at zero then declare the player or bot the winner or continue the fight accordingly.